requirements.py3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.py3.txt requirements.in
#
cffi==1.14.5
    # via cryptography
cryptography==3.4.7
    # via pyjwt
pycparser==2.20
    # via cffi
pyjwt[crypto]==1.7.1
    # via -r requirements.in
pytz==2021.1
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
six==1.14.0
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via -r requirements.in
typing==3.7.4.3
    # via -r requirements.in
